Update README.md
Proposed Amendment to add civility--People are largely negative about Kommittee, and feel conversation goes in circles of toxicity.  Calling the question has mostly worked in the past, but requires running through an (often 10+-person) queue in order to actually happen, meaning we wind up going in a lot of circles we don't need to.  Allowing the chair (with the help of someone else) to spot and disarm these cycles before they get going will reduce the time we spend pointlessly butting heads.