Update README.md
This is designed to avoid taking Kommittee time to "discuss our feelings about things that have been said/have happened."  One instance of this is the "kitchen allocation/heads-up wording mobpile SNAFU" of last semester.  The only time other than that which I've seen Kommittee time taken to conduct this sort of airing of our feelings was the Great Bias Incident of Spring 2011, which went similarly well, and spawned a number of other amendments. (V. H. and V. I. in particular were created out of that incident.)  While people may very well need to discuss their feelings and air them, doing so at a public meeting, where everyone is required to remain in the room at risk of being considered absent (Parlimentary Rules 9), and when doing such does not meaningfully contribute to the allocation of funds and resources for which Kommittee exists, is inappropriate.  As it stands, however, we have no way to prevent this, because without a motion, the question cannot be called.